requirements.txt
numpy==2.1.3
pandas==2.2.3
requests==2.32.3
PyYAML==6.0.2
python-dotenv==1.0.1
matplotlib==3.9.2
scikit-learn==1.5.2
